% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Close.R
\name{Close}
\alias{Close}
\title{Close}
\usage{
Close(query_obj)
}
\arguments{
\item{query_obj}{QueryClass instance-ID}
}
\value{
This function returns a list with the following items:
    \itemize{
      \item {info} {Info}
      \item {success} {A boolean, indicating if the command was completed successfull}
    }
}
\description{
Closes and unregisters the query with the specified ID
}
\details{
This method returns \emph{self} invisibly, thus making it possible to chain together multiple method calls.
}
