% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIntercept.R
\name{GetIntercept}
\alias{GetIntercept}
\title{GetIntercept}
\usage{
GetIntercept(session)
}
\arguments{
\item{session}{BasexClient instance-ID}
}
\value{
Current value
}
\description{
Current value for session$Intercept
}
