% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSuccess.R
\name{GetSuccess}
\alias{GetSuccess}
\title{GetSuccess}
\usage{
GetSuccess(session)
}
\arguments{
\item{session}{BasexClient instance-ID}
}
\value{
Current value
}
\description{
Current value from session$Success
}
