% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Info.R
\name{Info}
\alias{Info}
\title{Info}
\usage{
Info(query_obj)
}
\arguments{
\item{query_obj}{QueryClass instance-ID}
}
\description{
Returns a string with query compilation and profiling info.
}
\details{
If the query object has not been executed yet, an empty string is returned.
}
