% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_to_raw.R
\name{input_to_raw}
\alias{input_to_raw}
\title{input_to_raw}
\usage{
input_to_raw(input)
}
\arguments{
\item{input}{Character vector length 1}
}
\value{
'Raw' vector
}
\description{
Convert \emph{input} to a length-1 character vector.
}
\details{
If \emph{input} is a reference to a file, the number of bytes
    corresponding to the size is read. If it is an URL, the URL is read and converted to a 'Raw' vector.
    The function does not catch errors.
}
