% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewBaseXClient.R
\name{NewBasexClient}
\alias{NewBasexClient}
\title{Title}
\usage{
NewBasexClient(host = "localhost", port = 1984, user, password)
}
\arguments{
\item{host, port}{Host name and port-number}

\item{user, password}{User credentials}
}
\value{
BasexClient-instance
}
\description{
Create a BaseX-client
}
\details{
This creates a BaseX-client. By default it listens to port 1984 on localhost.
    Username and password should be changed after the installation of 'BaseX'.
}
\examples{

\dontrun{
session <- NewBasexClient(user = <username>, password = "<password>")
}

}
