% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Options.R
\name{Options}
\alias{Options}
\title{Options}
\usage{
Options(query_obj)
}
\arguments{
\item{query_obj}{QueryClass instance-ID}
}
\description{
Returns a string with all query serialization parameters, which
    can be assigned to the serializer option.
}
\details{
For a list of possibe types see \url{http://docs.basex.org/wiki/Java_Bindings#Data_Types}
}
