% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.R
\name{put}
\alias{put}
\title{Put}
\usage{
put(session, path, input)
}
\arguments{
\item{session}{BasexClient instance-ID}

\item{path}{Path where to store the data}

\item{input}{Add or replacement}
}
\value{
A list with two items
    \itemize{
      \item {info} {Aditional info}
      \item {success} {A boolean, indicating if the command was completed successfull}
    }
}
\description{
Adds or replaces a resource with the specified input.
}
\details{
The input can be a UTF-8 encoded XML document, a binary resource, or any other data (such as JSON or CSV)
    that can be successfully converted to a resource by the server.
    This method returns \emph{self} invisibly, thus making it possible to chain together multiple method calls.
}
\examples{
\dontrun{
put(Session, "test", "<xml>Create test</xml>")
}

}
