% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{Query}
\alias{Query}
\title{Query}
\usage{
Query(session, query_string)
}
\arguments{
\item{session}{BasexClient instance-ID}

\item{query_string}{query string}
}
\value{
Query_ID
}
\description{
Creates a new query instance and returns it's id.
}
\details{
If paste0() is used to create a multi-line statement, the lines must be separeted by a space or a newline \\n-character.
}
\examples{
\dontrun{
query_txt <- "for $i in 1 to 2 return <xml>Text { $i }</xml>"
query_obj <- Query(Session, query_txt)
print(Execute(query_obj))
}

}
