% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestoreIntercept.R
\name{RestoreIntercept}
\alias{RestoreIntercept}
\title{RestoreIntercept}
\usage{
RestoreIntercept(session)
}
\arguments{
\item{session}{BasexClient instance-ID}
}
\description{
Restore Intercept to original new value
}
\details{
This method returns \emph{self} invisibly, thus making it possible to chain together multiple method calls.
}
