% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result2matrix.R
\name{result2matrix}
\alias{result2matrix}
\title{result2matrix}
\usage{
result2matrix(input, cols)
}
\arguments{
\item{input}{Query-result}

\item{cols}{Number of columns}
}
\value{
Return result from query as matrix
}
\description{
Converts the query-result to a matrix. The query-result is a list.
    'cols' is needed to determine the number of columns.
}
