% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result2tibble.R
\name{result2tibble}
\alias{result2tibble}
\title{result2tibble}
\usage{
result2tibble(...)
}
\arguments{
\item{...}{Query-result}
}
\value{
Return result from query as tibble
}
\description{
Converts the query-result to a tibble. The query-result is either a list (sequence) or an array.
    If it is a list, 'cols' is needed to determine the number of columns.
}
