% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetIntercept.R
\name{SetIntercept}
\alias{SetIntercept}
\title{SetIntercept}
\usage{
SetIntercept(session, intercept)
}
\arguments{
\item{session}{BasexClient instance-ID}

\item{intercept}{New Intercept value}
}
\description{
Assign a new value to session$Intercept
}
\details{
This method returns \emph{self} invisibly, thus making it possible to chain together multiple method calls.
}
\examples{
\dontrun{
SetIntercept(TRUE)
}

}
