% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetSuccess.R
\name{SetSuccess}
\alias{SetSuccess}
\title{SetSuccess}
\usage{
SetSuccess(session, success)
}
\arguments{
\item{session}{BasexClient instance-ID}

\item{success}{Success-indicator for the last operation on the socket}
}
\description{
Assign a new value to session$Success
}
\examples{
\dontrun{
SetSuccess(TRUE)
}

}
