% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SocketClass.R
\name{SocketClass}
\alias{SocketClass}
\title{SocketClass}
\description{
All methods that are used by BasexClient and QueryClass
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SocketClass$new()}}
\item \href{#method-finalize}{\code{SocketClass$finalize()}}
\item \href{#method-handShake}{\code{SocketClass$handShake()}}
\item \href{#method-write_Byte}{\code{SocketClass$write_Byte()}}
\item \href{#method-clone}{\code{SocketClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new socket
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SocketClass$new(host, port = 1984L, username, password)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{host, port, username, password}}{Host-information and credentials}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
When releasing the session-object, close the socketConnection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SocketClass$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-handShake"></a>}}
\if{latex}{\out{\hypertarget{method-handShake}{}}}
\subsection{Method \code{handShake()}}{
Send input to the socket and return the response
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SocketClass$handShake(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{Input}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Input is a raw vector, built up by converting all input to raw and concatenating the results
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_Byte"></a>}}
\if{latex}{\out{\hypertarget{method-write_Byte}{}}}
\subsection{Method \code{write_Byte()}}{
Write 1 byte to the socket
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SocketClass$write_Byte(Byte)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Byte}}{A  vector length 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SocketClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
