% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Updating.R
\name{Updating}
\alias{Updating}
\title{Updating}
\usage{
Updating(query_obj)
}
\arguments{
\item{query_obj}{Query instance-ID}
}
\value{
This function returns a list with the following items:
    \itemize{
      \item {result} {Result}
      \item {success} {A boolean, indicating if the command was completed successfull}
    }
}
\description{
Check if the query contains updating expressions.
}
\details{
Returns \emph{TRUE} if the query contains updating expressions;
    \emph{FALSE} otherwise.
}
