% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add.R
\name{Add}
\alias{Add}
\title{Add}
\usage{
Add(session, path, input)
}
\arguments{
\item{session}{BasexClient instance-ID}

\item{path}{Path}

\item{input}{Additional input (optional).}
}
\value{
A list with two items
    \itemize{
      \item {info} {Aditional info}
      \item {success} {Boolean, indicating if the command was completed successfull}
    }
}
\description{
Adds a new resource to the opened database.
}
\details{
The 'input' can be a length-1 character vector or a stream.
    The utility-function \emph{input_to_raw} can be used to convert
    a file or an URL to a stream.
}
\examples{
\dontrun{
Add(Session, "test", "<xml>Add</xml>")
}

}
