% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Info.R
\name{Info}
\alias{Info}
\title{Info}
\usage{
Info(query_obj)
}
\arguments{
\item{query_obj}{QueryClass instance-ID}
}
\value{
This function returns a list with the following items:
    \itemize{
      \item {Info} {Info}
      \item {success} {A boolean, indicating if the command was completed successfull}
    }
}
\description{
Returns a string with query compilation and profiling info.
}
\details{
If the query object has not been executed yet, an empty string is returned.
}
