% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/More.R
\name{More}
\alias{More}
\title{More}
\usage{
More(query_obj)
}
\arguments{
\item{query_obj}{QueryClass instance-ID}
}
\value{
Boolean
}
\description{
Indicates if there are any other results in the query-result.
}
\examples{
\dontrun{
query2 <- "for $i in 1 to 3 return <xml>Text { $i }</xml>"
query_iterate <- Query(Session, query2)
while (More(query_iterate)) {
  cat(Next(query_iterate), "\n")
  }
}

}
