% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Next.R
\name{Next}
\alias{Next}
\title{Next}
\usage{
Next(query_obj)
}
\arguments{
\item{query_obj}{QueryClass instance-ID}
}
\description{
Returns the next result when iterating over a query
}
\examples{
\dontrun{
query2 <- "for $i in 1 to 3 return <xml>Text { $i }</xml>"
query_iterate <- Query(Session, query2)
while (More(query_iterate)) {
  cat(Next(query_iterate), "\n")
  }
}

}
