% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace.R
\name{Replace}
\alias{Replace}
\title{Replace}
\usage{
Replace(session, path, input)
}
\arguments{
\item{session}{BasexClient instance-ID}

\item{path}{Path where to store the data}

\item{input}{Replacement}
}
\value{
A list with two items
    \itemize{
      \item {info} {Aditional info}
      \item {success} {A boolean, indicating if the command was completed successfull}
    }
}
\description{
Replaces a resource with the specified input.
}
\details{
This method returns \emph{self} invisibly, thus making it possible to chain together multiple method calls.
}
\examples{
\dontrun{
Replace(Session, "test", "<xml>Create test</xml>")
}

}
