% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result2tibble.R
\name{result2tibble}
\alias{result2tibble}
\title{result2tibble}
\usage{
result2tibble(input, cols)
}
\arguments{
\item{input}{Query-result}

\item{cols}{Number of columns}
}
\value{
Return result from query as tibble
}
\description{
Converts the query-result to a tibble. The query-result is a list.
    'cols' is needed to determine the number of columns.
}
