% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Updating.R
\name{Updating}
\alias{Updating}
\title{Updating}
\usage{
Updating(query_obj)
}
\arguments{
\item{query_obj}{Query instance-ID}
}
\value{
Boolean
}
\description{
Check if the query contains updating expressions.
}
\details{
Returns \emph{TRUE} if the query contains updating expressions;
    \emph{FALSE} otherwise.
}
